+++ app/http/middleware/refresh_token.go
@@ -16,7 +16,7 @@ func (RefreshTokenMiddleware) Handle(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		var a auth.JWTAuth
 		conf := foundation.RetrieveConfig()
-		a.RefreshToken(w, conf.App.Key)
+		a.RefreshToken(w, conf.Key)
 
 		next.ServeHTTP(w, r)
 	})