+++ config/redis.go
@@ -1,0 +1,14 @@
+package config
+
+type redisConf struct {
+	Host     string
+	Port     int
+	User     string
+	Password string
+}
+
+func GetRedis() *redisConf {
+	return &redisConf{
+		Port: 6379,
+	}
+}