+++ goweb.go
@@ -3,8 +3,11 @@ package main
 import (
 	"encoding/gob"
 	"github.com/RobyFerro/go-web-framework"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/RobyFerro/go-web/database/model"
 	"github.com/RobyFerro/go-web/register"
+	"github.com/joho/godotenv"
+	"log"
 )
 
 // Main Go-Web entry point.
@@ -12,5 +15,9 @@ func main() {
 	gob.Register(&model.User{})
 
 	entities := register.BaseEntities()
-	foundation.Start(entities)
+	if err := godotenv.Load(".env"); err != nil {
+		log.Fatalln(err)
+	}
+
+	foundation.Start(entities, config.GetSever())
 }