+++ service/db_connection.go
@@ -2,15 +2,15 @@ package service
 
 import (
 	"fmt"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 	"github.com/labstack/gommon/log"
 )
 
 // ConnectDB to sql database
-func ConnectDB(conf *app.Conf) *gorm.DB {
-	connectionString, driver := createConnectionString(conf)
+func ConnectDB() *gorm.DB {
+	connectionString, driver := createConnectionString()
 	db, err := gorm.Open(driver, connectionString)
 
 	if err != nil {
@@ -21,13 +21,14 @@ func ConnectDB(conf *app.Conf) *gorm.DB {
 }
 
 // Create string for SQL connection
-func createConnectionString(conf *app.Conf) (string, string) {
+func createConnectionString() (string, string) {
+	conf := config.GetSQL()
 	return fmt.Sprintf(
 		"%s:%s@(%s:%d)/%s?charset=utf8&parseTime=True&loc=Local",
-		conf.Database.User,
-		conf.Database.Password,
-		conf.Database.Host,
-		conf.Database.Port,
-		conf.Database.Database,
-	), conf.Database.Driver
+		conf.User,
+		conf.Password,
+		conf.Host,
+		conf.Port,
+		conf.Database,
+	), conf.Driver
 }