+++ service/elasticsearch.go
@@ -1,15 +1,16 @@
 package service
 
 import (
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/labstack/gommon/log"
 )
 
 // ConnectElastic returns an ElasticSearch client
-func ConnectElastic(conf *app.Conf) *elasticsearch.Client {
+func ConnectElastic() *elasticsearch.Client {
+	elasticConf := config.GetElastic()
 	cfg := elasticsearch.Config{
-		Addresses: conf.Elastic.Hosts,
+		Addresses: elasticConf.Hosts,
 	}
 	es, err := elasticsearch.NewClient(cfg)