+++ service/mongo.go
@@ -3,7 +3,7 @@ package service
 import (
 	"context"
 	"fmt"
-	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/config"
 	"github.com/labstack/gommon/log"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -11,8 +11,10 @@ import (
 )
 
 // ConnectMongo returns a connection to MongoDB
-func ConnectMongo(conf *app.Conf) *mongo.Database {
-	client, err := mongo.NewClient(options.Client().ApplyURI(fmt.Sprintf("mongodb://%s:%d", conf.Mongo.Host, conf.Mongo.Port)))
+func ConnectMongo() *mongo.Database {
+	conf := config.GetMongo()
+	conn := fmt.Sprintf("mongodb://%s:%d", conf.Host, conf.Port)
+	client, err := mongo.NewClient(options.Client().ApplyURI(conn))
 
 	if err != nil {
 		log.Error(err)
@@ -21,5 +23,5 @@ func ConnectMongo(conf *app.Conf) *mongo.Database {
 	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
 	_ = client.Connect(ctx)
 
-	return client.Database(conf.Mongo.Database)
+	return client.Database(conf.Database)
 }